INT01
git-svn-id: svn://svn.code.sf.net/p/rosecheckers/code/trunk@40 a2ac0d72-06d8-44ab-9e58-0f2374626789

+++ c_rules/ncce/EXP34_C.c
@@ -1,7 +1,7 @@
 #include <string.h>
 #include <stdlib.h>
 
-const int SIZE_MAX = 50;
+const size_t SIZE_MAX = 50;
 
 void EXP34_C() {
   char* input_str = "Hi, there!";
